requirements.txt
# Core Dependencies
python-dotenv==1.0.1
websockets==14.1
requests==2.32.3
eth-account==0.13.4

# Hyperliquid SDK
hyperliquid-python-sdk==0.9.0

